+++ Changelog.md
@@ -7,6 +7,7 @@ NovusCodeLibrary4.SimpleTemplate
 * Updated TagName without <%xxx%> save in HashTable for key
 * New function public string CleanTagName(string aRawTagEx)
 * New function public void UpdateTagValue(string aTagName, string aTagValue )
+* New property IgnoreBlankTags - If TagValue is blank re-insert <%TagName%>